backend/requirements.txt
# 智能数据洞察助手 - 后端依赖

# 核心依赖
Flask>=3.0.0
Flask-CORS>=4.0.0

# LangChain 生态
langchain>=1.2.12
langchain-core>=0.2.29
langchain-community>=0.2.19
langchain-openai>=0.1.23
langchain-anthropic>=0.1.12

# LangGraph
langgraph>=1.1.1

# 向量数据库
chromadb>=1.5.5
faiss-cpu>=1.7.4

# 文本嵌入模型
sentence-transformers>=2.2.2

# LLM API客户端
openai>=1.35.12
anthropic>=0.35.0

# 工具依赖
tiktoken>=0.7.0
pydantic>=2.7.1
python-dotenv>=1.0.0

# 数据处理
numpy>=1.26.4
pandas>=2.2.2

# 文档处理
PyPDF2>=3.0.1

# 其他
tqdm>=4.66.4

# 开发依赖
pytest>=7.4.3
black>=24.4.2
flake8>=6.1.0
